+++ coralogix/data_source_coralogix_webhook_test.go
@@ -1,11 +1,11 @@
 // Copyright 2024 Coralogix Ltd.
-// 
+//
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     https://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -30,19 +30,34 @@ func TestAccCoralogixDataSourceWebhook_basic(t *testing.T) {
 		Steps: []resource.TestStep{
 			{
 				Config: testAccCoralogixResourceSlackWebhook(w) +
-					testAccCoralogixDataSourceWebhook_read(),
+					testAccCoralogixDataSourceWebhook_readByID(),
 				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr("data.coralogix_webhook.test", "name", w.name),
-					resource.TestCheckResourceAttr("data.coralogix_webhook.test", "slack.url", w.url),
+					resource.TestCheckResourceAttr("data.coralogix_webhook.test_by_id", "name", w.name),
+					resource.TestCheckResourceAttr("data.coralogix_webhook.test_by_id", "slack.url", w.url),
+				),
+			},
+			{
+				Config: testAccCoralogixResourceSlackWebhook(w) +
+					testAccCoralogixDataSourceWebhook_readByName(),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr("data.coralogix_webhook.test_by_name", "name", w.name),
+					resource.TestCheckResourceAttr("data.coralogix_webhook.test_by_name", "slack.url", w.url),
 				),
 			},
 		},
 	})
 }
 
-func testAccCoralogixDataSourceWebhook_read() string {
-	return `data "coralogix_webhook" "test" {
+func testAccCoralogixDataSourceWebhook_readByID() string {
+	return `data "coralogix_webhook" "test_by_id" {
 	id = coralogix_webhook.test.id
+}
+`
+}
+
+func testAccCoralogixDataSourceWebhook_readByName() string {
+	return `data "coralogix_webhook" "test_by_name" {
+	name = coralogix_webhook.test.name
 }
 `
 }